Remove unnecessary sampling and race conditions in tests
Also change how spans are fetched. This seems to be the only way to do it
safely without turning on syncPush in tracer which I can't think of a good way
to do using these high level public API tests.

Basically there were race conditions of finishing a trace which enqueued into
the push queue and flushing the traces.