Fixed sarama instrumentation and introduced ContextExt interface
Sarama instrumentation was identifying spans with a non-unique and
unelibale key which caused a number of issues with span reporting as
some spans were not being finished. This commit replaces the said key
with the span ID.

This span ID is extracted from the kafka message headers and used to
identify any pending spans. This requires Kafka (not sarama) version
0.11.0 or newer which was released in 2017.